test/inspector/regress/regress-crbug-980018-expected.txt

Regression test for crbug.com/980018

Running test: testClassNames
{
    result : {
        className : A
        description : A(0)
        objectId : <objectId>
        subtype : map
        type : object
    }
}
{
    result : {
        className : B
        description : B(0)
        objectId : <objectId>
        subtype : map
        type : object
    }
}
{
    result : {
        className : C
        description : C(0)
        objectId : <objectId>
        subtype : map
        type : object
    }
}
